Implemented Classbrowser's openInNewChildWindowOf
The main goal is to be able to open applications in new windows but the controller runs from the initial window (the Launcher).
This is essential for live editing.
Without this changes to the source code should first be published to the server before another window can use the new source.